refactor: replacing auth with auth 2
