+++ cognite_toolkit/_cdf_tk/apps/_core_app.py
@@ -16,7 +16,7 @@ from cognite_toolkit._cdf_tk.commands import BuildCommand, CleanCommand, DeployC
 from cognite_toolkit._cdf_tk.commands.clean import AVAILABLE_DATA_TYPES
 from cognite_toolkit._cdf_tk.exceptions import ToolkitFileNotFoundError
 from cognite_toolkit._cdf_tk.utils import get_cicd_environment
-from cognite_toolkit._cdf_tk.utils.auth2 import EnvironmentVariables
+from cognite_toolkit._cdf_tk.utils.auth import EnvironmentVariables
 from cognite_toolkit._version import __version__ as current_version
