+++ cognite_toolkit/_cdf_tk/utils/auth.py
@@ -1,24 +1,10 @@
-# Copyright 2023 Cognite AS
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-from __future__ import annotations
-
+import itertools
 import os
-from dataclasses import _MISSING_TYPE, dataclass, field, fields
-from typing import Any, Literal, TypeAlias
+from collections.abc import Iterable, Mapping
+from dataclasses import Field, dataclass, field, fields
+from typing import Any, Literal, TypeAlias, get_args
 
-import typer
-from cognite.client.config import global_config
+import questionary
 from cognite.client.credentials import (
     CredentialProvider,
     OAuthClientCredentials,
@@ -26,24 +12,19 @@ from cognite.client.credentials import (
     OAuthInteractive,
     Token,
 )
-from cognite.client.data_classes import ClientCredentials
-from cognite.client.exceptions import CogniteAPIError
+from questionary import Choice
 from rich import print
 
 from cognite_toolkit._cdf_tk.client import ToolkitClient, ToolkitClientConfig
-from cognite_toolkit._cdf_tk.constants import (
-    _RUNNING_IN_BROWSER,
-    TOOLKIT_CLIENT_ENTRA_ID,
-)
-from cognite_toolkit._cdf_tk.exceptions import (
-    AuthenticationError,
-    ToolkitValueError,
-)
-from cognite_toolkit._cdf_tk.tk_warnings import IgnoredValueWarning
+from cognite_toolkit._cdf_tk.constants import TOOLKIT_CLIENT_ENTRA_ID
+from cognite_toolkit._cdf_tk.exceptions import AuthenticationError, ToolkitKeyError, ToolkitMissingValueError
+from cognite_toolkit._cdf_tk.utils import humanize_collection
 from cognite_toolkit._version import __version__
 
 LoginFlow: TypeAlias = Literal["client_credentials", "token", "device_code", "interactive"]
 Provider: TypeAlias = Literal["entra_id", "cdf", "other"]
+VALID_LOGIN_FLOWS = get_args(LoginFlow)
+VALID_PROVIDERS = get_args(Provider)
 
 CLIENT_NAME = f"CDF-Toolkit:{__version__}"
 LOGIN_FLOW_DESCRIPTION = {
@@ -52,8 +33,7 @@ LOGIN_FLOW_DESCRIPTION = {
     "device_code": "Login using the browser with your user credentials using device code flow",
     "token": "Use a Token directly to authenticate",
 }
-
-PROVDER_DESCRIPTION = {
+PROVIDER_DESCRIPTION = {
     "entra_id": "Use Microsoft Entra ID to authenticate",
     "cdf": "Use Cognite IDP to authenticate",
     "other": "Use other IDP to authenticate",
@@ -61,530 +41,420 @@ PROVDER_DESCRIPTION = {
 
 
 @dataclass
-class AuthVariables:
-    cluster: str | None = field(
-        metadata=dict(env_name="CDF_CLUSTER", display_name="CDF cluster", example="westeurope-1")
-    )
-    project: str | None = field(metadata=dict(env_name="CDF_PROJECT", display_name="CDF project", example="publicdata"))
-    cdf_url: str | None = field(
+class EnvOptions(Mapping):
+    display_name: str
+    example: str
+    is_secret: bool = False
+    required: frozenset[tuple[LoginFlow, Provider | None]] = frozenset()
+    optional: frozenset[tuple[LoginFlow, Provider | None]] = frozenset()
+
+    def __getitem__(self, key: str) -> str | bool:
+        return self.__dict__[key]
+
+    def __iter__(self) -> Iterable[str]:  # type: ignore[override]
+        return iter(self.__dict__.keys())
+
+    def __len__(self) -> int:
+        return len(self.__dict__)
+
+
+ALL_CASES = [(flow, None) for flow in VALID_LOGIN_FLOWS]
+
+
+@dataclass
+class EnvironmentVariables:
+    CDF_CLUSTER: str = field(metadata=EnvOptions("CDF cluster", "westeurope-1"))
+    CDF_PROJECT: str = field(metadata=EnvOptions("CDF project", "publicdata"))
+    LOGIN_FLOW: LoginFlow = field(default="client_credentials", metadata=EnvOptions("Login flow", "client_credentials"))
+    PROVIDER: Provider = field(default="entra_id", metadata=EnvOptions("Provider", "entra_id"))
+    CDF_URL: str | None = field(
         default=None,
-        metadata=dict(env_name="CDF_URL", display_name="CDF URL", example="https://CDF_CLUSTER.cognitedata.com"),
+        metadata=EnvOptions("CDF URL", "https://{CDF_CLUSTER}.cognitedata.com", optional=frozenset(ALL_CASES)),
     )
-    login_flow: LoginFlow = field(
-        default="client_credentials",
-        metadata=dict(
-            env_name="LOGIN_FLOW",
-            display_name="Login flow",
-            example="client_credentials",
+    CDF_TOKEN: str | None = field(
+        default=None, metadata=EnvOptions("OAuth2 token", example="", required=frozenset([("token", None)]))
+    )
+    IDP_CLIENT_ID: str | None = field(
+        default=None,
+        metadata=EnvOptions(
+            display_name="client id",
+            example="",
+            required=frozenset([("client_credentials", None), ("interactive", None), ("device_code", "other")]),
         ),
     )
-    provider: Provider = field(
-        default="entra_id",
-        metadata=dict(
-            env_name="PROVIDER",
-            display_name="Provider",
-            example="entra_id",
+    IDP_CLIENT_SECRET: str | None = field(
+        default=None,
+        metadata=EnvOptions(
+            display_name="client secret", example="", is_secret=True, required=frozenset([("client_credentials", None)])
         ),
     )
-    token: str | None = field(
-        default=None, metadata=dict(env_name="CDF_TOKEN", display_name="OAuth2 token", example="")
-    )
-    client_id: str | None = field(
+    IDP_TOKEN_URL: str | None = field(
         default=None,
-        metadata=dict(
-            env_name="IDP_CLIENT_ID", display_name="client id", example="XXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX"
-        ),
-    )
-    client_secret: str | None = field(
-        default=None,
-        metadata=dict(env_name="IDP_CLIENT_SECRET", display_name="client secret", example=""),
-    )
-    token_url: str | None = field(
-        default=None,
-        metadata=dict(
-            env_name="IDP_TOKEN_URL",
+        metadata=EnvOptions(
             display_name="token URL",
-            example="https://login.microsoftonline.com/IDP_TENANT_ID/oauth2/v2.0/token",
+            example="https://login.microsoftonline.com/{IDP_TENANT_ID}/oauth2/v2.0/token",
+            required=frozenset([("client_credentials", "other"), ("interactive", None)]),
+            optional=frozenset([("client_credentials", "entra_id")]),
         ),
     )
-    tenant_id: str | None = field(
+    IDP_TENANT_ID: str | None = field(
         default=None,
-        metadata=dict(
-            env_name="IDP_TENANT_ID",
+        metadata=EnvOptions(
             display_name="Tenant id for MS Entra",
-            example="12345678-1234-1234-1234-123456789012",
+            example="mytenant.onmicrosoft.com",
+            required=frozenset([("device_code", "entra_id"), ("client_credentials", "entra_id")]),
         ),
     )
-    audience: str | None = field(
+    IDP_AUDIENCE: str | None = field(
         default=None,
-        metadata=dict(
-            env_name="IDP_AUDIENCE",
+        metadata=EnvOptions(
             display_name="IDP audience",
-            example="https://CDF_CLUSTER.cognitedata.com",
+            example="https://{CDF_CLUSTER}.cognitedata.com",
+            optional=frozenset([("client_credentials", "entra_id"), ("client_credentials", "other")]),
         ),
     )
-    scopes: str | None = field(
+
+    IDP_SCOPES: str | None = field(
         default=None,
-        metadata=dict(
-            env_name="IDP_SCOPES",
+        metadata=EnvOptions(
             display_name="IDP scopes",
-            example="https://CDF_CLUSTER.cognitedata.com/.default",
+            example="https://{CDF_CLUSTER}.cognitedata.com/.default",
+            optional=frozenset(
+                [("client_credentials", "entra_id"), ("client_credentials", "other"), ("interactive", None)]
+            ),
         ),
     )
-    authority_url: str | None = field(
+    IDP_AUTHORITY_URL: str | None = field(
         default=None,
-        metadata=dict(
-            env_name="IDP_AUTHORITY_URL",
+        metadata=EnvOptions(
             display_name="IDP authority URL",
-            example="https://login.microsoftonline.com/IDP_TENANT_ID",
+            example="https://login.microsoftonline.com/{IDP_TENANT_ID}",
+            optional=frozenset([("interactive", None)]),
         ),
     )
-    oidc_discovery_url: str | None = field(
+    IDP_DISCOVERY_URL: str | None = field(
         default=None,
-        metadata=dict(
-            env_name="IDP_DISCOVERY_URL",
+        metadata=EnvOptions(
             display_name="IDP OIDC discovery URL (root URL excl. /.well-known/...)",
             example="https://<auth0-tenant>.auth0.com/oauth",
+            required=frozenset([("device_code", "other")]),
         ),
     )
-    cdf_client_timeout: int = field(
-        default=30,
-        metadata=dict(
-            env_name="CDF_CLIENT_TIMEOUT",
-            display_name="CDF client timeout",
-            example="30",
-        ),
+    CDF_CLIENT_TIMEOUT: int = field(
+        default=30, metadata=EnvOptions(display_name="CDF client timeout", example="30", optional=frozenset(ALL_CASES))
     )
-    cdf_client_max_workers: int = field(
+    CDF_CLIENT_MAX_WORKERS: int = field(
         default=5,
-        metadata=dict(
-            env_name="CDF_CLIENT_MAX_WORKERS",
-            display_name="CDF client max workers",
-            example="5",
-        ),
+        metadata=EnvOptions(display_name="CDF client max workers", example="5", optional=frozenset(ALL_CASES)),
     )
+    _client: ToolkitClient | None = field(init=False, default=None)
 
     def __post_init__(self) -> None:
-        # Set defaults based on cluster and tenant_id
-        if self.cluster:
-            self.set_cluster_defaults()
-        if self.provider == "cdf":
-            self.set_cdf_provider_defaults()
-        if self.tenant_id and self.provider != "cdf":
-            self.set_token_id_defaults()
-        if self.token and self.login_flow != "token":
-            print(
-                f"  [bold yellow]Warning[/] CDF_TOKEN detected. This will override LOGIN_FLOW, "
-                f"thus LOGIN_FLOW={self.login_flow} will be ignored"
-            )
-            self.login_flow = "token"
+        self.LOGIN_FLOW = self.LOGIN_FLOW.lower()  # type: ignore[assignment]
+        if self.LOGIN_FLOW not in VALID_LOGIN_FLOWS:
+            raise AuthenticationError(f"Invalid login flow: {self.LOGIN_FLOW}. Valid options are {VALID_LOGIN_FLOWS}")
 
-        if isinstance(self.cdf_client_timeout, str):
-            try:
-                self.cdf_client_timeout = int(self.cdf_client_timeout)
-            except ValueError:
-                raise ToolkitValueError(f"CDF_CLIENT_TIMEOUT must be an integer, got {self.cdf_client_timeout}")
-        if isinstance(self.cdf_client_max_workers, str):
-            try:
-                self.cdf_client_max_workers = int(self.cdf_client_max_workers)
-            except ValueError:
-                raise ToolkitValueError(f"CDF_CLIENT_MAX_WORKERS must be an integer, got {self.cdf_client_max_workers}")
-
-    def set_cdf_provider_defaults(self, force: bool = False) -> None:
-        default_url = "https://auth.cognite.com/oauth2/token"
-        if force:
-            self.token_url = default_url
-        else:
-            self.token_url = self.token_url or default_url
-        if self.scopes is not None:
-            IgnoredValueWarning("IDP_SCOPES", self.scopes, "Provider Cog-IDP does not need scopes").print_warning()
-        self.scopes = None
-
-    def set_token_id_defaults(self) -> None:
-        if self.tenant_id:
-            self.token_url = self.token_url or f"https://login.microsoftonline.com/{self.tenant_id}/oauth2/v2.0/token"
-            self.authority_url = self.authority_url or f"https://login.microsoftonline.com/{self.tenant_id}"
-
-    def set_cluster_defaults(self) -> None:
-        if self.cluster:
-            self.cdf_url = self.cdf_url or f"https://{self.cluster}.cognitedata.com"
-            self.audience = self.audience or f"https://{self.cluster}.cognitedata.com"
-            if self.provider != "cdf":
-                self.scopes = self.scopes or f"https://{self.cluster}.cognitedata.com/.default"
+    # All derived properties
+    @property
+    def idp_token_url(self) -> str:
+        if self.PROVIDER == "cdf":
+            return "https://auth.cognite.com/oauth2/token"
+        if self.IDP_TOKEN_URL:
+            return self.IDP_TOKEN_URL
+        if self.PROVIDER == "entra_id" and self.IDP_TENANT_ID:
+            return f"https://login.microsoftonline.com/{self.IDP_TENANT_ID}/oauth2/v2.0/token"
+        alternative = ""
+        if self.PROVIDER == "entra_id":
+            alternative = " or provide IDP_TENANT_ID"
+        raise ToolkitKeyError(
+            f"IDP_TOKEN_URL is missing. Please provide it{alternative} in the environment variables.",
+            "IDP_TOKEN_URL",
+        )
 
     @property
-    def is_complete(self) -> bool:
-        if self.cdf_url is None:
-            return False
-        if self.login_flow == "token":
-            return self.token is not None
-        elif self.login_flow == "interactive":
-            return self.client_id is not None and self.tenant_id is not None and self.scopes is not None
-        elif self.login_flow == "client_credentials" and self.provider == "cdf":
-            return self.client_id is not None and self.client_secret is not None
-        elif self.login_flow == "client_credentials":
-            return (
-                self.client_id is not None
-                and self.client_secret is not None
-                and self.token_url is not None
-                and self.scopes is not None
-                and self.audience is not None
-            )
-        elif self.login_flow == "device_code" and self.provider == "entra_id":
-            return self.tenant_id is not None
-        elif self.login_flow == "device_code" and self.provider == "other":
-            return self.client_id is not None and self.oidc_discovery_url is not None
-        return False
+    def cdf_url(self) -> str:
+        return self.CDF_URL or f"https://{self.CDF_CLUSTER}.cognitedata.com"
+
+    @property
+    def idp_audience(self) -> str:
+        return self.IDP_AUDIENCE or f"https://{self.CDF_CLUSTER}.cognitedata.com"
+
+    @property
+    def idp_scopes(self) -> list[str]:
+        if self.IDP_SCOPES:
+            return self.IDP_SCOPES.split(",")
+        return [f"https://{self.CDF_CLUSTER}.cognitedata.com/.default"]
+
+    @property
+    def idp_authority_url(self) -> str:
+        if self.IDP_AUTHORITY_URL:
+            return self.IDP_AUTHORITY_URL
+        if self.PROVIDER == "entra_id" and self.IDP_TENANT_ID:
+            return f"https://login.microsoftonline.com/{self.IDP_TENANT_ID}"
+        alternative = ""
+        if self.PROVIDER == "entra_id":
+            alternative = " or provide IDP_TENANT_ID"
+        raise ToolkitKeyError(
+            f"IDP_AUTHORITY_URL is missing. Please provide it{alternative} in the environment variables.",
+            "IDP_AUTHORITY_URL",
+        )
 
     @classmethod
-    def from_env(cls, override: dict[str, Any] | None = None) -> AuthVariables:
-        override = override or {}
-        args: dict[str, Any] = {}
-        for field_ in fields(cls):
-            if env_name := field_.metadata.get("env_name"):
-                default = None if isinstance(field_.default, _MISSING_TYPE) else field_.default
-                args[field_.name] = override.get(env_name, os.environ.get(env_name, default))
+    def _fields(cls, inst: "EnvironmentVariables | None" = None) -> tuple[Field, ...]:
+        return tuple(f for f in fields(inst or cls) if not f.name.startswith("_"))
+
+    @classmethod
+    def create_from_environment(cls) -> "EnvironmentVariables":
+        if missing := [key for key in ["CDF_CLUSTER", "CDF_PROJECT"] if key not in os.environ]:
+            raise ToolkitMissingValueError(f"Missing environment variables: {humanize_collection(missing)}")
+        args: dict[str, Any] = {
+            field_.name: field_.type(os.environ[field_.name]) if field_.type is int else os.environ[field_.name]  # type: ignore[operator]
+            for field_ in cls._fields()
+            if field_.name in os.environ
+        }
         return cls(**args)
 
-    def create_dotenv_file(self) -> str:
-        lines = [
-            "# .env file generated by cognite-toolkit",
-            self._write_var("login_flow"),
-            self._write_var("cluster"),
-            self._write_var("project"),
-        ]
-        if self.login_flow != "token":
-            lines += [
-                self._write_var("provider"),
-            ]
+    def get_credentials(self) -> CredentialProvider:
+        method_by_flow = {
+            "client_credentials": self._get_oauth_client_credentials,
+            "interactive": self._get_oauth_interactive,
+            "device_code": self._get_oauth_device_code,
+            "token": self._get_token,
+        }
+        if self.LOGIN_FLOW not in method_by_flow:
+            # Should already be checked in __post_init__
+            raise AuthenticationError(f"Login flow {self.LOGIN_FLOW} is not supported.")
 
-        if self.login_flow == "token":
-            lines += [
-                "# When using a token, the IDP variables are not needed, so they are not included.",
-                self._write_var("token"),
-            ]
-        elif self.login_flow == "client_credentials":
-            lines += [
-                self._write_var("client_id"),
-                self._write_var("client_secret"),
-            ]
-        elif self.login_flow == "device_code" and self.provider != "entra_id":
-            ...
-        elif self.login_flow == "interactive":
-            lines += [
-                self._write_var("client_id"),
-            ]
-        else:
-            raise ValueError(f"Login flow {self.login_flow} is not supported.")
-        if self.login_flow in ("client_credentials", "interactive") and self.provider != "cdf":
-            lines += [
-                "# Note: Either the TENANT_ID or the TENANT_URL must be written.",
-                self._write_var("tenant_id"),
-                self._write_var("token_url"),
-            ]
-        elif self.login_flow == "client_credentials" and self.provider == "cdf":
-            lines += [
-                self._write_var("token_url"),
-            ]
-        elif self.login_flow == "device_code" and self.provider == "entra_id":
-            lines += [
-                self._write_var("tenant_id"),
-            ]
-        elif self.login_flow == "device_code" and self.provider == "other":
-            lines += [
-                self._write_var("client_id"),
-                self._write_var("oidc_discovery_url"),
-            ]
-        lines += [
-            "# The below variables are the defaults, they are automatically constructed unless they are set.",
-            self._write_var("cdf_url"),
-        ]
-        if self.login_flow in ("client_credentials", "interactive") and self.provider != "cdf":
-            lines += [
-                self._write_var("scopes"),
-            ]
-        if self.login_flow == "interactive":
-            lines += [
-                self._write_var("authority_url"),
-            ]
-        if (self.login_flow == "client_credentials" and self.provider != "cdf") or (
-            self.login_flow == "device_code" and self.provider == "other"
-        ):
-            lines += [
-                self._write_var("audience"),
-            ]
-        lines += [
-            "# The below variables control the client configuration.",
-            self._write_var("cdf_client_timeout"),
-            self._write_var("cdf_client_max_workers"),
-        ]
+        if missing_vars := self.get_missing_vars():
+            raise ToolkitMissingValueError(
+                f"The login flow '{self.LOGIN_FLOW}' requires the following environment variables: {humanize_collection(missing_vars)}.",
+            )
+        return method_by_flow[self.LOGIN_FLOW]()
 
-        return "\n".join(lines)
+    def _get_oauth_client_credentials(self) -> OAuthClientCredentials:
+        if self.PROVIDER == "cdf":
+            return OAuthClientCredentials(
+                client_id=self.IDP_CLIENT_ID,  # type: ignore[arg-type]
+                client_secret=self.IDP_CLIENT_SECRET,  # type: ignore[arg-type]
+                token_url=self.idp_token_url,
+                scopes=None,  # type: ignore[arg-type]
+            )
+        return OAuthClientCredentials(
+            client_id=self.IDP_CLIENT_ID,  # type: ignore[arg-type]
+            client_secret=self.IDP_CLIENT_SECRET,  # type: ignore[arg-type]
+            token_url=self.idp_token_url,
+            audience=self.idp_audience,
+            scopes=self.idp_scopes,
+        )
 
-    def _write_var(self, var_name: str) -> str:
-        value = getattr(self, var_name)
-        field_ = _auth_field_by_name[var_name].metadata
-        if value is None:
-            return f"{field_['env_name']}="
-        return f"{field_['env_name']}={value}"
+    def _get_oauth_interactive(self) -> OAuthInteractive:
+        return OAuthInteractive(
+            client_id=self.IDP_CLIENT_ID,  # type: ignore[arg-type]
+            authority_url=self.idp_authority_url,
+            scopes=self.idp_scopes,
+        )
 
-
-_auth_field_by_name = {field.name: field for field in fields(AuthVariables)}
-
-
-class CDFToolConfig:
-    """Configurations for how to store data in CDF
-
-    Properties:
-        toolkit_client: active ToolkitClient
-
-    """
-
-    def __init__(
-        self,
-        token: str | None = None,
-        cluster: str | None = None,
-        project: str | None = None,
-        cdf_url: str | None = None,
-        auth_vars: AuthVariables | None = None,
-        skip_initialization: bool = False,
-    ) -> None:
-        self._environ: dict[str, str | None] = {}
-        # If cluster, project, or token are passed as arguments, we override the environment variables.
-        # This means these will be used when we initialize the CogniteClient when we initialize from
-        # environment variables. Note if we are running in the browser, we will not use these arguments.
-        if project:
-            self._environ["CDF_PROJECT"] = project
-        if cluster:
-            self._environ["CDF_CLUSTER"] = cluster
-        if token:
-            self._environ["CDF_TOKEN"] = token
-        if cdf_url:
-            self._environ["CDF_URL"] = cdf_url
-
-        # ClientName is used for logging usage of the CDF-Toolkit.
-        self._client_name = CLIENT_NAME
-
-        self._cluster: str | None = cluster
-        self._project: str | None = project
-        self._cdf_url: str | None = None
-        self._scopes: list[str] = []
-        self._audience: str | None = None
-        self._token_url: str | None = None
-        self._credentials_args: dict[str, Any] = {}
-        self._credentials_provider: CredentialProvider | None = None
-        self._toolkit_client: ToolkitClient | None = None
-
-        global_config.disable_pypi_version_check = True
-        global_config.silence_feature_preview_warnings = True
-        if _RUNNING_IN_BROWSER and auth_vars is None:
-            self._initialize_in_browser()
-            return
-
-        self._auth_vars = auth_vars or AuthVariables.from_env(self._environ)
-        if not skip_initialization:
-            self.initialize_from_auth_variables(self._auth_vars)
-        self._login_flow = self._auth_vars.login_flow
-
-    def _initialize_in_browser(self) -> None:
-        try:
-            self._toolkit_client = ToolkitClient()
-        except CogniteAPIError as e:
-            raise AuthenticationError(f"Failed to initialize CogniteClient in browser: {e}")
-
-        if self._cluster or self._project:
-            print("[bold yellow]Warning[/] Cluster and project are arguments ignored when running in the browser.")
-        self._cluster = self._toolkit_client.config.base_url.removeprefix("https://").split(".", maxsplit=1)[0]
-        self._project = self._toolkit_client.config.project
-        self._cdf_url = self._toolkit_client.config.base_url
-
-    def initialize_from_auth_variables(self, auth: AuthVariables, clear_cache: bool = False) -> None:
-        """Initialize the CDFToolConfig from the AuthVariables and returns whether it was successful or not."""
-        cluster = auth.cluster or self._cluster
-        project = auth.project or self._project
-
-        if cluster is None or project is None:
-            raise AuthenticationError("Cluster and Project must be set to authenticate the client.")
-
-        self._cluster = cluster
-        self._project = project
-        self._cdf_url = auth.cdf_url or self._cdf_url
-
-        if auth.login_flow == "token":
-            if not auth.token:
-                raise AuthenticationError("Login flow=token is set but no CDF_TOKEN is not provided.")
-            self._credentials_args = dict(token=auth.token)
-            self._credentials_provider = Token(**self._credentials_args)
-        elif auth.login_flow == "device_code" and auth.provider == "entra_id":
+    def _get_oauth_device_code(self) -> OAuthDeviceCode:
+        if self.PROVIDER == "entra_id":
             # TODO: If the user has submitted the wrong scopes, we may get a valid token that gives 401 on the CDF API.
             # The user will then have to wait until the token has expired to retry with the correct scopes.
             # If we add clear_cache=True to the OAuthDeviceCode, the token cache will be cleared.
             # We could add a cli option to auth verify, e.g. --clear-token-cache, that will clear the cache.
-            if not auth.tenant_id:
-                raise ValueError("IDP_TENANT_ID is required for device code login.")
-            # For Entra ID, we have defaults for everything, even the app registration as we can use the CDF public app.
-            self._credentials_args = dict(
-                tenant_id=auth.tenant_id,
+            return OAuthDeviceCode.default_for_azure_ad(
+                tenant_id=self.IDP_TENANT_ID,  # type: ignore[arg-type]
                 client_id=TOOLKIT_CLIENT_ENTRA_ID,
-                cdf_cluster=self._cluster,
-                clear_cache=clear_cache,
+                cdf_cluster=self.CDF_CLUSTER,
+                clear_cache=False,
             )
-            self._credentials_provider = OAuthDeviceCode.default_for_azure_ad(**self._credentials_args)
-        elif auth.login_flow == "device_code" and auth.provider == "other":
-            if not auth.client_id:
-                raise ValueError("IDP_CLIENT_ID is required for device code login.")
-            self._credentials_args = dict(
+        elif self.PROVIDER == "other":
+            return OAuthDeviceCode(
                 authority_url=None,
-                cdf_cluster=auth.cluster,
-                oauth_discovery_url=auth.oidc_discovery_url,
-                client_id=auth.client_id,
-                audience=auth.audience,
-                clear_cache=clear_cache,
+                cdf_cluster=self.CDF_CLUSTER,
+                oauth_discovery_url=self.IDP_DISCOVERY_URL,
+                client_id=self.IDP_CLIENT_ID,  # type: ignore[arg-type]
+                audience=self.idp_audience,
             )
-            self._credentials_provider = OAuthDeviceCode(**self._credentials_args)
-        elif auth.login_flow == "interactive":
-            if auth.scopes:
-                self._scopes = auth.scopes.split(",")
-            if not (auth.client_id and auth.authority_url and auth.scopes):
-                raise AuthenticationError(
-                    "Login flow=interactive is set but missing required authentication "
-                    "variables: IDP_CLIENT_ID and IDP_TENANT_ID (or IDP_AUTHORITY_URL). Cannot authenticate the client."
-                )
-            self._credentials_args = dict(
-                authority_url=auth.authority_url,
-                client_id=auth.client_id,
-                scopes=self._scopes,
-            )
-            self._credentials_provider = OAuthInteractive(**self._credentials_args)
-        elif auth.login_flow == "client_credentials" and auth.provider == "cdf":
-            if not (auth.client_id and auth.client_secret):
-                raise AuthenticationError(
-                    "Login flow=client_credentials is set but missing required authentication "
-                    "variables: IDP_CLIENT_ID and IDP_CLIENT_SECRET. Cannot authenticate the client."
-                )
-            self._credentials_args = dict(
-                token_url=auth.token_url,
-                client_id=auth.client_id,
-                client_secret=auth.client_secret,
-                scopes=None,
-            )
-            self._credentials_provider = OAuthClientCredentials(**self._credentials_args)
-        elif auth.login_flow == "client_credentials" or auth.login_flow is None:
-            if auth.login_flow is None:
-                print(
-                    "  [bold yellow]Warning[/] No login flow is set. Defaulting to client_credentials. "
-                    "Set LOGIN_FLOW to 'client_credentials', 'token', or 'interactive' to avoid this warning."
-                )
-            if auth.scopes:
-                self._scopes = auth.scopes.split(",")
-            if auth.audience:
-                self._audience = auth.audience
-
-            if not (auth.token_url and auth.client_id and auth.client_secret and self._scopes and self._audience):
-                raise AuthenticationError(
-                    "Login flow=client_credentials is set but missing required authentication "
-                    "variables: IDP_CLIENT_ID, IDP_CLIENT_SECRET and IDP_TENANT_ID (or IDP_TOKEN_URL). "
-                    "Cannot authenticate the client."
-                )
-            self._credentials_args = dict(
-                token_url=auth.token_url,
-                client_id=auth.client_id,
-                client_secret=auth.client_secret,
-                scopes=self._scopes,
-                audience=self._audience,
-            )
-            self._credentials_provider = OAuthClientCredentials(**self._credentials_args)
         else:
-            raise AuthenticationError(f"Login flow {auth.login_flow} is not supported.")
-        self._token_url = auth.token_url
-        self._toolkit_client = ToolkitClient(
-            ToolkitClientConfig(
-                client_name=self._client_name,
-                base_url=self._cdf_url,
-                project=self._project,
-                credentials=self._credentials_provider,
-                timeout=auth.cdf_client_timeout,
-            )
+            raise AuthenticationError(f"The provider {self.PROVIDER} is not supported for device code flow.")
+
+    def _get_token(self) -> Token:
+        if not self.CDF_TOKEN:
+            raise ToolkitKeyError("CDF_TOKEN must be set in the environment", "CDF_TOKEN")
+        return Token(self.CDF_TOKEN)
+
+    def get_config(self) -> ToolkitClientConfig:
+        return ToolkitClientConfig(
+            client_name=CLIENT_NAME,
+            project=self.CDF_PROJECT,
+            credentials=self.get_credentials(),
+            base_url=self.cdf_url,
+            timeout=self.CDF_CLIENT_TIMEOUT,
+            max_workers=self.CDF_CLIENT_MAX_WORKERS,
         )
-        global_config.max_workers = auth.cdf_client_max_workers
-        self._update_environment_variables()
-        self._auth_vars = auth
 
-    def _update_environment_variables(self) -> None:
-        """This updates the cache environment variables with the auth
-        variables.
+    def get_client(self) -> ToolkitClient:
+        if self._client is None:
+            self._client = ToolkitClient(config=self.get_config())
+        return self._client
 
-        This is necessary for the .as_string() method to dump correctly.
-        """
-        for field_name in ["cluster", "project", "cdf_url", "scopes", "audience"]:
-            try:
-                field_ = _auth_field_by_name[field_name]
-                env_name = field_.metadata["env_name"]
-                value = getattr(self, f"_{field_name}")
-            except KeyError as e:
-                # This means that the attribute is not set correctly in AuthVariables,
-                # ensure that 'env_name' is set in the metadata for all fields in AuthVariables.
-                raise RuntimeError("AuthVariables not created correctly. Contact Support") from e
-
-            if value:
-                self._environ[env_name] = value[0] if isinstance(value, list) else value
-
-    @classmethod
-    def from_context(cls, ctx: typer.Context) -> CDFToolConfig:
-        if ctx.obj.mockToolGlobals is not None:
-            return ctx.obj.mockToolGlobals
-        else:
-            return CDFToolConfig()
-
-    def environment_variables(self) -> dict[str, str | None]:
-        return {**self._environ.copy(), **os.environ}
+    def dump(self, include_os: bool = True) -> dict[str, str | None]:
+        variables: dict[str, Any] = {}
+        for field_ in self._fields(self):
+            default_value = field_.name.casefold()
+            if hasattr(self, default_value):
+                value = getattr(self, default_value)
+            else:
+                value = getattr(self, field_.name)
+            if isinstance(value, list):
+                value = ",".join(value)
+            if value is not None:
+                variables[field_.name] = str(value)
+        if include_os:
+            variables.update(os.environ)
+        return variables
 
     def as_string(self) -> str:
-        environment = self._environ.copy()
-        if "IDP_CLIENT_SECRET" in environment:
-            environment["IDP_CLIENT_SECRET"] = "***"
-        if "TRANSFORMATIONS_CLIENT_SECRET" in environment:
-            environment["TRANSFORMATIONS_CLIENT_SECRET"] = "***"
-        envs = ""
-        for e in environment:
-            envs += f"  {e}={environment[e]}\n"
-        return f"CDF Project {self._project} in cluster {self._cluster}:\n{envs}"
+        env_lines: list[str] = []
+        for field_, value in itertools.chain(self.get_required_with_value(), self.get_optional_with_value()):
+            if value is None:
+                continue
+            if isinstance(value, list):
+                value = ",".join(value)
+            if field_.metadata["is_secret"]:
+                value = "*" * 5
+            env_lines.append(f"  {field_.name}={value}")
 
-    @property
-    def toolkit_client(self) -> ToolkitClient:
-        if self._toolkit_client is None:
-            raise ValueError("ToolkitClient is not initialized.")
-        return self._toolkit_client
+        body = "\n".join(env_lines)
+        return f"CDF Project {self.CDF_PROJECT} in cluster {self.CDF_CLUSTER}:\n{body}"
 
-    @property
-    def project(self) -> str:
-        if self._project is None:
-            raise ValueError("Project is not initialized.")
-        return self._project
+    def get_missing_vars(self) -> set[str]:
+        flow, provider = self.LOGIN_FLOW, self.PROVIDER
+        missing: set[str] = set()
+        for field_ in self._fields(self):
+            required = field_.metadata["required"]
+            value = getattr(self, field_.name)
+            if value is None and required and ((flow, provider) in required or (flow, None) in required):
+                missing.add(field_.name)
 
-    @property
-    def cdf_cluster(self) -> str:
-        if self._cluster is None:
-            raise ValueError("Cluster is not initialized.")
-        return self._cluster
+        if provider == "entra_id" and "IDP_TOKEN_URL" in missing:
+            missing -= {"IDP_TOKEN_URL"}
+            if self.IDP_TENANT_ID is None:
+                missing.add("IDP_TENANT_ID")
+        return missing
 
-    def create_client(self, credentials: ClientCredentials) -> ToolkitClient:
-        if self._auth_vars.token_url is None or self._auth_vars.scopes is None:
-            raise AuthenticationError("Token URL and Scopes are required to create a client.")
-        if self._project is None or self._cdf_url is None:
-            raise AuthenticationError("Project and CDF URL are required to create a client.")
-        service_credentials = OAuthClientCredentials(
-            token_url=self._auth_vars.token_url,
-            client_id=credentials.client_id,
-            client_secret=credentials.client_secret,
-            scopes=self._auth_vars.scopes.split(",") if self._auth_vars.scopes else [],
-        )
+    def get_required_with_value(self) -> list[tuple[Field, Any]]:
+        flow, provider = self.LOGIN_FLOW, self.PROVIDER
+        values: list[tuple[Field, Any]] = []
+        for field_ in self._fields(self):
+            required = field_.metadata["required"]
+            if required and ((flow, provider) in required or (flow, None) in required):
+                if field_.name == "IDP_TOKEN_URL" and provider == "entra_id":
+                    continue
+                values.append((field_, getattr(self, field_.name)))
+        return values
 
-        return ToolkitClient(
-            config=ToolkitClientConfig(
-                client_name=self._client_name,
-                project=self._project,
-                base_url=self._cdf_url,
-                credentials=service_credentials,
-            )
-        )
+    def get_optional_with_value(self) -> list[tuple[Field, Any]]:
+        flow, provider = self.LOGIN_FLOW, self.PROVIDER
+        values: list[tuple[Field, Any]] = []
+        for field_ in self._fields(self):
+            optional = field_.metadata["optional"]
+            if optional and ((flow, provider) in optional or (flow, None) in optional):
+                if (default_name := field_.name.casefold()) and hasattr(self, default_name):
+                    value = getattr(self, default_name)
+                else:
+                    value = getattr(self, field_.name)
+                values.append((field_, value))
+        return values
+
+    def create_dotenv_file(self) -> str:
+        lines = [
+            "# .env file generated by cognite-toolkit",
+            f"CDF_CLUSTER={self.CDF_CLUSTER}",
+            f"CDF_PROJECT={self.CDF_PROJECT}",
+            f"LOGIN_FLOW={self.LOGIN_FLOW}",
+        ]
+        if self.LOGIN_FLOW != "token":
+            lines += [
+                f"PROVIDER={self.PROVIDER}",
+            ]
+        lines.append("")
+        lines.append("# Required variables")
+        for field_, value in self.get_required_with_value():
+            if value is not None:
+                lines.append(f"{field_.name}={value}")
+        lines.append("")
+        lines.append("# Optional variables (derived from the required variables)")
+        for field_, value in self.get_optional_with_value():
+            if value is None:
+                continue
+            if isinstance(value, list):
+                value = ",".join(value)
+            lines.append(f"{field_.name}={value}")
+        return "\n".join(lines) + "\n"
+
+
+def prompt_user_environment_variables(current: EnvironmentVariables | None = None) -> EnvironmentVariables:
+    login_flow = questionary.select(
+        "Choose the login flow",
+        choices=[
+            Choice(title=f"{flow}: {description}", value=flow) for flow, description in LOGIN_FLOW_DESCRIPTION.items()
+        ],
+        default=current.LOGIN_FLOW if current else "client_credentials",
+    ).ask()
+    provider = questionary.select(
+        "Choose the provider",
+        choices=[
+            Choice(title=f"{provider}: {description}", value=provider)
+            for provider, description in PROVIDER_DESCRIPTION.items()
+        ],
+        default=current.PROVIDER if current else "entra_id",
+    ).ask()
+    cdf_cluster = questionary.text("Enter the CDF cluster", default=current.CDF_CLUSTER if current else "").ask()
+    cdf_project = questionary.text("Enter the CDF project", default=current.CDF_PROJECT if current else "").ask()
+    args: dict[str, Any] = current.dump(include_os=False) if current else {}
+    args.update(
+        {"LOGIN_FLOW": login_flow, "CDF_CLUSTER": cdf_cluster, "CDF_PROJECT": cdf_project, "PROVIDER": provider}
+    )
+    env_vars = EnvironmentVariables(**args)
+    idp_tenant_id = env_vars.IDP_TENANT_ID or "IDP_TENANT_ID"
+    for field_, value in env_vars.get_required_with_value():
+        user_value = get_user_value(field_, value, cdf_cluster, idp_tenant_id)
+        setattr(env_vars, field_.name, user_value)
+        if field_.name == "IDP_TENANT_ID":
+            idp_tenant_id = user_value
+
+    optional_values = env_vars.get_optional_with_value()
+    for field_, value in optional_values:
+        print(f"  {field_.name}={value}")
+    if questionary.confirm("Do you want to change any of these variables?", default=False).ask():
+        for field_, value in optional_values:
+            user_value = get_user_value(field_, value, cdf_cluster, idp_tenant_id)
+            setattr(env_vars, field_.name, user_value)
+    return env_vars
+
+
+def get_user_value(field_: Field, value: Any, cdf_cluster: str, idp_tenant_id: str) -> Any:
+    is_secret = field_.metadata["is_secret"]
+    display_name = field_.metadata["display_name"]
+    default = value or field_.metadata["example"].format(CDF_CLUSTER=cdf_cluster, IDP_TENANT_ID=idp_tenant_id)
+    if isinstance(value, list):
+        default = ",".join(value)
+    elif value is not None and not isinstance(value, str):
+        default = str(value)
+    if is_secret:
+        user_value = questionary.password(f"Enter the {display_name}:", default=default).ask()
+    else:
+        user_value = questionary.text(f"Enter the {display_name}:", default=default).ask()
+    if field_.type is int:
+        try:
+            user_value = int(user_value)
+        except ValueError:
+            print(f"Invalid value: {user_value}. Please enter an integer.")
+            return get_user_value(field_, value, cdf_cluster, idp_tenant_id)
+    return user_value
+
+
+if __name__ == "__main__":
+    # For easy testing
+    envs = prompt_user_environment_variables()
+    print(envs)
