+++ tests/test_unit/test_cdf_tk/test_loaders/test_data_model_loader.py
@@ -10,7 +10,7 @@ from cognite_toolkit._cdf_tk.client.data_classes.graphql_data_models import Grap
 from cognite_toolkit._cdf_tk.exceptions import ToolkitCycleError
 from cognite_toolkit._cdf_tk.loaders import DataModelLoader, ResourceWorker
 from cognite_toolkit._cdf_tk.loaders._resource_loaders import GraphQLLoader
-from cognite_toolkit._cdf_tk.utils.auth2 import EnvironmentVariables
+from cognite_toolkit._cdf_tk.utils.auth import EnvironmentVariables
 from tests.test_unit.approval_client import ApprovalToolkitClient
