+++ tests/test_unit/test_cdf_tk/test_loaders/test_data_set_loader.py
@@ -1,7 +1,7 @@
 from cognite.client.data_classes import DataSet
 
 from cognite_toolkit._cdf_tk.loaders import DataSetsLoader, ResourceWorker
-from cognite_toolkit._cdf_tk.utils.auth2 import EnvironmentVariables
+from cognite_toolkit._cdf_tk.utils.auth import EnvironmentVariables
 from tests.data import LOAD_DATA
 from tests.test_unit.approval_client import ApprovalToolkitClient
