+++ tests/test_unit/test_cdf_tk/test_loaders/test_extraction_pipeline_loader.py
@@ -18,7 +18,7 @@ from cognite_toolkit._cdf_tk.loaders import (
     ResourceLoader,
     ResourceWorker,
 )
-from cognite_toolkit._cdf_tk.utils.auth2 import EnvironmentVariables
+from cognite_toolkit._cdf_tk.utils.auth import EnvironmentVariables
 from tests.test_unit.approval_client import ApprovalToolkitClient
