+++ tests/test_unit/test_cdf_tk/test_loaders/test_function_loader.py
@@ -5,7 +5,7 @@ from cognite.client.data_classes import Function, FunctionWrite
 
 from cognite_toolkit._cdf_tk.loaders import FunctionLoader, ResourceWorker
 from cognite_toolkit._cdf_tk.utils import calculate_directory_hash, calculate_secure_hash
-from cognite_toolkit._cdf_tk.utils.auth2 import EnvironmentVariables
+from cognite_toolkit._cdf_tk.utils.auth import EnvironmentVariables
 from tests.data import LOAD_DATA
 from tests.test_unit.approval_client import ApprovalToolkitClient
